# src/main/resources/openapi/openapi.yml
openapi: 3.0.1
info:
  title: Сервис "Виртуальная школа"
  version: v0.r1
servers:
  - url: http://localhost:8080
    description: Сервер приложения Virtual-School
paths:
  /homework/{id}:
    get:
      tags:
        - loaded-homework
      operationId: getCompletedHomework
      parameters:
        - name: id
          in: path
          required: true
          description: id домашнего задания
          example: 1
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadedHomeworkDto'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
...
  /homework:
    post:
      tags:
        - loaded-homework
      operationId: sendCompletedHomework
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadedHomeworkDto'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
...
  /student/change-snils:
    put:
      description: Изменения данных СНИЛС
      tags:
        - student
      operationId: changeSnilsStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentEntity'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
  /student/change-birthcertificate:
    put:
      description: Изменение данных свидетельства о рождении
      tags:
        - student
      operationId: changeBirthCertificateStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentEntity'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
  /student/add-passport:
    put:
      description: добавление паспорта ученику достигшему возраста 14 лет
      tags:
        - student
      operationId: addPassportStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
        required: true
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
  /student:
    description: Добавление нового ученика в список школы
    post:
      tags:
        - student
      operationId: addStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentEntity'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - orderAuth: [ ]
...
components:
  securitySchemes:
    orderAuth:
      type: http
      scheme: basic
  responses:
    UnauthorizedError:
      description: Аутентификация не пройдена
      headers:
        WWW_Authenticate:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: Неверные входные данные
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Ошибка сервера.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Заказ не найден
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    BirthCertificateDto:
      type: object
      description: ДТО свидетельства о рождении
      properties:
        serial:
          type: string
          example: 1234
        number:
          type: string
          example: 567890
    PassportDto:
      type: object
      description: ДТО паспотра
      properties:
        serial:
          type: string
          example: 1234
        number:
          type: string
          example: 567890
    SnilsDto:
      type: object
      description: ДТО СНИЛСа
      properties:
        numberView:
          type: string
          example: 1234567890
    StudentDto:
      description: ДТО студента
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Ivan
        snils:
          $ref: '#/components/schemas/SnilsDto'
        birthCertificate:
          $ref: '#/components/schemas/BirthCertificateDto'
        passport:
          $ref: '#/components/schemas/PassportDto'
        birthday:
          type: string
          format: date
    StudentEntity:
      type: object
      description: сущность студента, которая будет сохраняться в базу
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Ivan
        snils:
          type: string
          example: 1234567890
        birthCertificate:
          type: string
        passport:
          type: string
          example: 1234
        birthday:
          type: string
          format: date
          example: 567890
    LoadedHomeworkDto:
      type: object
      description: ДТО решенного учеником домашнего задания
      properties:
        homeworkId:
          type: integer
          format: int64
          example: 1
        studentId:
          type: integer
          format: int64
          example: 1
        taskCompletionResult:
          type: string
          example: any text
        homeworkForClassId:
          type: integer
          format: int64
          example: 1
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |
            A relative URI reference that uniquely identifies the problem type.
          default: 'about:blank'
          example: '/problem/internal-server-error'
        title:
          type: string
          description: |
            title of the problem
          example: Internal server error
        status:
          type: integer
          format: int32
          description: |
            The HTTP status of the problem.
          minimum: 100
          maximum: 599
          example: 404
        detail:
          type: string
          description: |
            Detail of the problem
          example: Save data error